requirements.txt
# Core Flask Framework
Flask==3.0.3
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
flasgger==0.9.7.1

# Web Server
gunicorn==23.0.0
Werkzeug==3.0.4

# Database
SQLAlchemy==2.0.36
alembic==1.14.0

# Authentication & Security
python-dotenv==1.1.1

# AI/ML & LLM
openai==2.2.0

# HTTP & Requests
requests==2.32.5
httpx==0.28.1

# Environment & Configuration
click==8.1.8

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0

# Development Tools
black==24.10.0
isort==5.13.2
flake8==7.1.1

# Utilities
PyYAML==6.0.3
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
